refactor(agent): split loop.go by concern layer into separate files
- Extract LLM retry logic to llm_retry.go (318 lines)
- Extract tool execution logic to tool_exec.go (380 lines)
- Extract runtime telemetry to runtime_meta.go (239 lines)
- Extract LLM streaming logic to llm_stream.go (542 lines)
- loop.go reduced from 1981 to 571 lines (71% reduction)
- All functions stay in package agent, no import cycles
- Full test suite passes: go test ./... -count=1
