+++ pkg/agent/result.go
