+++ pkg/agent/tool_guard.go
