+++ pkg/agent/tool_guard_test.go
